Add support for renaming drives
This adds a new config option '[[drive]]' that currently has a
'name' and 'newname' field. If a drive would be created that has
the name 'name' (e.g., Drive1), then it will be shown with the
new name instead. This will also change the names given to the
default partition when creating a new drive. It will not change
the partition name for an already existing drive (though it will
change the way the names are shown on the pISO itself).

This will also not work with multiple levels of snapshot. So
if 'Drive1' is renamed to 'Pictures' and backed up twice, then
the pISO will show 'Drive1-backup-backup'. If this becomes a
problem, it shouldn't be too difficult to fix.