Fix bug where discs would not mount is MacOS (fixes #32)
This was caused by MacOS issuing a READ DVD STRUCTURE command to determine
the type of disc being read. When we failed to respond to this, the OS
starts to attempt to read with CD READ requests instead of standard READ
requests. We don't support CD READ currently. Now we respond to the first
request correctly, so MacOS does normal READs and things work correctly.